requirements.txt
streamlit
opencv-python-headless
matplotlib
plotly

soundfile
imageio
spacy
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz

transformers==4.29.0
datasets
huggingface_hub

accelerate
diffusers
sentencepiece
torchaudio